[verified] fix: critical C2(pom test scope) + C3(windows guard) + W6(volume overflow x2) + W7(timezone) + W8(try..catch x2) + W9(saveAll batch)

+++ src/main/java/com/stock/invest/service/impl/TigerStockServiceImpl.java
@@ -201,7 +201,7 @@ public class TigerStockServiceImpl {
             StockInfo stockInfo = new StockInfo();
             stockInfo.setSymbol(symbol);
             stockInfo.setCurrentPrice(latest.getClose());
-            stockInfo.setVolume((int)latest.getVolume());
+            stockInfo.setVolume(latest.getVolume());
             stockInfo.setChange(latest.getClose() - latest.getOpen());
             stockInfo.setChangePercent((latest.getClose() - latest.getOpen()) / latest.getOpen() * 100);
